[libc++][AIX] Add config for building with the libcxx CI infrastructure
This change adds the AIX configuration to run-buildbot, an AIX CMake cache file,
and appropriate compiler and linker flags for testing AIX to the
lit configuration files. Either of the 32-bit or 64-bit configurations
can be built by setting `OBJECT_MODE` in the build environment (as is
typical for AIX).